refactor(mcp_server): split tools.rs (1438 lines) into tools/ directory
Extracted into four focused sub-modules under mcp_server/tools/:
- types.rs: McpToolSpec, ToolCallError, shared constants
- specs.rs: tool spec builders and JSON schema helpers
- params.rs: build_rpc_params and all argument validation helpers
- dispatch.rs: call_tool, list_tools_result, agent/subagent handlers

External API (tool_specs, McpToolSpec, call_tool, list_tools_result,
tool_success, tool_error) preserved via mod.rs re-exports. Companion
tools_tests.rs included via #[path] with test-gated pub use re-exports
so use super::* continues to work unchanged.
