+++ src/openhuman/agent/task_dispatcher/mod.rs
@@ -1,0 +1,40 @@
+//! Deterministic task-card dispatcher.
+//!
+//! Turns a [`TaskBoardCard`] into work: it **claims** the card via a
+//! compare-and-set (re-load the board and transition only a `Todo`/`Ready`
+//! card to `in_progress`, so a stale/concurrent re-dispatch of the same card
+//! is rejected), runs a single **autonomous agent turn** toward the card's
+//! objective, and **writes the outcome back** to the board (`done` + evidence
+//! on success, `blocked` + reason on failure).
+//!
+//! This is the one executor both dispatch paths converge on:
+//! - the **board poller** (cards that arrived without a proactive trigger), and
+//! - the **proactive triage** arm (`agent::triage::apply_decision`), once it has
+//!   decided to act on a task-board card.
+//!
+//! The runner mirrors `skills::spawn_workflow_run_background`: build the
+//! `orchestrator` agent fresh inside a detached task, cap tool iterations, and
+//! run `agent.run_single` under `with_autonomous_iter_cap`. PR-4 generalises the
+//! executor from the default agent to a resolved personality/skill; this module
+//! keeps the default-agent path so the pipeline runs end-to-end first.
+
+mod dispatch;
+mod executor;
+mod poller;
+mod prompt;
+mod registry;
+mod types;
+
+#[cfg(test)]
+mod tests;
+
+// ── Public API ────────────────────────────────────────────────────────────────
+
+pub use dispatch::dispatch_card;
+pub use prompt::build_task_prompt;
+pub use registry::cancel_session;
+pub use poller::start_board_poller;
+pub use types::DispatchOutcome;
+
+// `pub(crate)` for test drivers.
+pub(crate) use poller::poll_once;
