+++ src/openhuman/meet_agent/brain/speech.rs
@@ -1,0 +1,68 @@
+//! STT and TTS adapters (real cloud paths).
+
+use base64::{engine::general_purpose::STANDARD as B64, Engine as _};
+use serde_json::json;
+
+use super::constants::{SAMPLE_RATE_HZ, TTS_MODEL_ID};
+use crate::openhuman::meet_agent::wav;
+
+// ─── Real STT adapter ───────────────────────────────────────────────
+
+pub(super) async fn stt(samples: &[i16]) -> Result<String, String> {
+    use crate::openhuman::voice::cloud_transcribe::{transcribe_cloud, CloudTranscribeOptions};
+
+    let config = crate::openhuman::config::ops::load_config_with_timeout().await?;
+    let wav_bytes = wav::pack_pcm16le_mono_wav(samples, SAMPLE_RATE_HZ);
+    let audio_b64 = B64.encode(&wav_bytes);
+    let opts = CloudTranscribeOptions {
+        mime_type: Some("audio/wav".to_string()),
+        file_name: Some("meet-agent.wav".to_string()),
+        ..Default::default()
+    };
+    let outcome = transcribe_cloud(&config, &audio_b64, &opts).await?;
+    let text = outcome.value.text.clone();
+    Ok(text)
+}
+
+// ─── Real TTS adapter ───────────────────────────────────────────────
+
+pub(super) async fn tts(text: &str) -> Result<Vec<i16>, String> {
+    use crate::openhuman::voice::reply_speech::{synthesize_reply, ReplySpeechOptions};
+
+    let config = crate::openhuman::config::ops::load_config_with_timeout().await?;
+    // Tuned for live conversational speech, not narration:
+    //   stability 0.4 — leave room for prosody / inflection. Higher
+    //     values (>0.6) flatten the read into the "monotone audiobook"
+    //     timbre the previous default produced.
+    //   similarity_boost 0.75 — keep the chosen voice's character.
+    //   style 0.35 — light expressiveness; too high makes punctuation
+    //     swallow words.
+    //   use_speaker_boost on — louder, clearer in noisy meetings.
+    let voice_settings = json!({
+        "stability": 0.4,
+        "similarity_boost": 0.75,
+        "style": 0.35,
+        "use_speaker_boost": true,
+    });
+    let opts = ReplySpeechOptions {
+        // Ask ElevenLabs (via the hosted backend) for raw PCM16LE @
+        // 16 kHz so we can feed the result straight into the
+        // shell-side bridge with no transcoding.
+        output_format: Some("pcm_16000".to_string()),
+        model_id: Some(TTS_MODEL_ID.to_string()),
+        voice_settings: Some(voice_settings),
+        ..Default::default()
+    };
+    let outcome = synthesize_reply(&config, text, &opts).await?;
+    let result = outcome.value;
+    let pcm_bytes = B64
+        .decode(result.audio_base64.as_bytes())
+        .map_err(|e| format!("decode tts base64: {e}"))?;
+    if !pcm_bytes.len().is_multiple_of(2) {
+        return Err(format!("odd byte length from tts: {}", pcm_bytes.len()));
+    }
+    Ok(pcm_bytes
+        .chunks_exact(2)
+        .map(|c| i16::from_le_bytes([c[0], c[1]]))
+        .collect())
+}
