+++ src/openhuman/meet_agent/brain_tests.rs
@@ -1,5 +1,7 @@
 use super::*;
 use crate::openhuman::meet_agent::session::registry;
+use crate::openhuman::meet_agent::types::{SessionEvent, SessionEventKind};
+use serde_json::json;
 
 #[tokio::test]
 async fn run_turn_skips_short_buffers() {
